feat(dashboard): tabbed dashboard with routed models, gear-configs, and submissions tabs
Consolidate contribution management under /dashboard using a daisyUI tab bar.
The parent dashboard.vue owns the hero, breadcrumb, and auth gate, then renders
routed child tabs via <NuxtPage>:

- /dashboard          -> redirects to /dashboard/models
- /dashboard/models   -> My Models (moved out of /models/mine)
- /dashboard/gear-configs
- /dashboard/submissions

/models/mine now redirects to /dashboard/models so old links keep working. The
gear-config and submission i18n was split into the respective child pages with
all 10 locales preserved; shell/auth keys plus new tabs.* keys live in the parent.

Post-submission CTAs on the contribute pages now point at /dashboard/submissions,
and the /models browse + upload wizard links point at /dashboard/models.
