+++ app/pages/contribute/registry.vue
@@ -209,7 +209,7 @@
                     <i class="fa-duotone fa-solid fa-plus-large mr-2"></i>
                     {{ t('success.submit_another') }}
                   </button>
-                  <NuxtLink to="/dashboard" class="btn btn-neutral btn-outline w-full">
+                  <NuxtLink to="/dashboard/submissions" class="btn btn-neutral btn-outline w-full">
                     {{ t('success.view_submissions') }}
                   </NuxtLink>
                 </div>
